ingestion/requirements.txt
# Ingestion Dependencies - Data Processing Pipeline

# Scientific data handling
xarray>=2023.1.0
numpy>=1.24.0
pandas>=2.0.0

# Database connectivity
psycopg2-binary>=2.9.0

# Machine learning for embeddings
sentence-transformers>=3.0.0

